Completed phase 2 of conversion to GObject (I think)
Phase 1 of conversion was the conversion of all STLWRT C source files to use
GObject in lieu of GtkObject -- GtkObject was a relic and broke compatibility
with GTK+ 3 and later; few programs still rely on GtkObject.  Phase 2 was the
conversion of STLWRT header files to use GObject.  Phase 3 (coming up) is
the attempted compilation of the main STLWRT library to see where to go from
here.