No more "fat" and "thin" classes
Even in GTK+ 3 and later, classes are packed full of function pointers.
This is much unlike the situation for instance structures, where every last
byte is shaved off the instance structure's size and most variables are
placed in the instance's private data area.  For this reason, I do not yet
see a reason to support two sizes of classes in the same way I support two
sizes of object instances.