No more "fat" and "thin" interfaces
Even in GTK+ 3 and later, interfaces are packed full of function pointers.
This is much unlike the situation for instance structures, where every last
byte is shaved off the instance structure's size and most variables are
placed in the instance's private data area.  For this reason, I do not yet
see a reason to support two sizes of interfaces in the same way I support two
sizes of object instances.  This commit is much like the previous commit, but
the previous commit removed fat and thin classes, not interfaces.