Moved instance private data declarations to after the instance properties
This was done so that the first element of the private data of a thin
STLWRT object is always the instance properties.  If the application
expects fat objects, then "public" instance properties are read from the
instance structure itself, like in GTK+ 2; if the application expects thin
objects, then the instance properties are addressed relative to the start
of the private data area of the instance.  Formerly (i.e. yesterday), the
private data declaration preceded the instance properties structure
declaration; however, in C a structure member may not be of a type which
has not yet been declared.  I thus had to move the private instance data
declarations below the instance properties structures so that I could add
a line to the top of each private data structure declaration:

  struct _GtkSomethingOrOtherInstanceProps  instance_properties;