Almost done converting icons to old style icons
The sole icons remaining which are the "new style" Tango-esque icons are
16/edit-clear.png, 16/gtk-select-color.png, and 24/edit-clear.png.  All
other icons are the classic style.  Some icons could not be copied from
the original GNOME / GTK icon themes because the icons simply did not
exist back then; for those I have had to transform, flip, scale, and
overlay images to create best approximations of the GNOME icon theme of
many years ago.  Good examples of such transformation exercises are
the "printer-paused" and similar images.