Removed PixBuf theme engine from source tree
The PixBuf theme engine was the last remaining theme engine whose source
code was available directly in the STLWRT source tree itself.  When the
rest of STLWRT is complete enough, the PixBuf theme engine and many other
classic theme engines will be ported to STLWRT specifically and bundled
in one package, probably named 'stlwrt-themes'; the PixBuf theme engine
is not going away permanently.  Every line of code is another line of
code to maintain, and when you're talking about tens of thousands of
lines to maintain, that's a lot of work when you're getting a project
off the ground.  I have thus decided to remove it from STLWRT.