Start of phase 3:  Fully STLWRT-ize every source file!
Now that the error log is only ~1.8 MB, now it's time to stop with the
random bugfix stuff and go head-in and convert all source files to use
the 'g*_get_props ()' functions instead of accessing instance structures
directly.  There's somewhere around 2,000 functions to convert, so I'll
be at it for a while!