Blasted through stage 3 for GDK part of STLWRT
Now what's left is the GTK part...  That's the hard part; since GTK+ 3 and
later don't expose very much (if any) of the instance structures of the
GDK to applications, we don't need any VTypes for GDK-type functions in
STLWRT.  The GTK portion, however, is a different matter.  Applications
subclass GTK objects (mostly widgets) all the time and fail if the size of
the instance data is not exactly correct.  However, the optimist that I
am, I realize that few if any applications subclass non-widget objects
such as GtkStyle, and even if they did, no valid GTK+ 3 application would
do so, because GtkStyle was deprecated and almost completely ignored from
GTK+ 3.0.0 onward.  So I'm still quite optimistic about the stuff,
nevertheless.

But I will use search and replace and even a few one-liners when I can!