Oops!  Fixed major typo in README
Man, how can anybody mess that one up?  Talk about a really stupid
function -- or programmer for that matter.  ;-)  The README originally
referenced the variable `child` as `widget` -- a complete faux pas
that would not have been ignored by the compiler!  One would tend to
wonder how many such errors are present in my code.  Oh well, we'll
see when we do a compilation process at the end of phase 3.
Also marked the README for posterity.