Removed some code rot from GtkCalendar source file; icons update
For the most part, the icons update was in preparation for the Windows
2000-esque file chooser coming soon -- a few larger icons were needed
for the left sidebar of the file chooser.  The icons were imported from
the GNOME 2.14 icon collection.  While at that, I changed the 'Open
Recently Used' icons to more clearly indicate that they open recently
used _files_.  The last step was also taken for consistency, since there
was not a special 48x48 icon for the icon I was _previously_ using to
show 'Open Recently Used'.