More source code cleanup
This commit serves to me as a milestone.  As I type this I am conceiving a
program which reads through the compiler's errors and fixes the errors
automatically.  However, I want to ensure that if the program goes haywire
I can simply type:

    git -C stlwrt checkout -f

...and I haven't lost any real work of mine.

(OK.  I'm sure there's a cleaner way to do this in Git, but I don't have the
patience right now to figure out what that cleaner way is.  If you want to
enlighten me, just send me mail; my address should be open for everyone to
see on these commits.)